ref: refactoring code
